+++ src/plugins/terrain/historic/inf20100227/generatorInfdev20100227.cpp
@@ -27,28 +27,28 @@ std::shared_ptr<Chunk> GeneratorInfdev20100227::GenerateChunk(Int2 chunkPos) {
 			int32_t regionZ = blockZ / 1024;
 			// Generate terrain height
 			float noiseGen1Value =
-				(float)(this->noiseGen1.get()->GenerateOctaves((double)((float)blockX / 0.03125F), 0.0,
-															   (double)((float)blockZ / 0.03125F)) -
-						this->noiseGen2.get()->GenerateOctaves((double)((float)blockX / 0.015625F), 0.0,
-															   (double)((float)blockZ / 0.015625F))) /
+				float(this->noiseGen1.get()->GenerateOctaves(double((float)blockX / 0.03125F), 0.0,
+															   double((float)blockZ / 0.03125F)) -
+						this->noiseGen2.get()->GenerateOctaves(double((float)blockX / 0.015625F), 0.0,
+															   double((float)blockZ / 0.015625F))) /
 				512.0F / 4.0F;
-			float noiseGen5Value = (float)this->noiseGen5.get()->GenerateOctaves((double)((float)blockX / 4.0F),
-																				 (double)((float)blockZ / 4.0F));
-			float noiseGen6Value = (float)this->noiseGen6.get()->GenerateOctaves((double)((float)blockX / 8.0F),
-																				 (double)((float)blockZ / 8.0F)) /
+			float noiseGen5Value = (float)this->noiseGen5.get()->GenerateOctaves(double((float)blockX / 4.0F),
+																				 double((float)blockZ / 4.0F));
+			float noiseGen6Value = (float)this->noiseGen6.get()->GenerateOctaves(double((float)blockX / 8.0F),
+																				 double((float)blockZ / 8.0F)) /
 								   8.0F;
 			noiseGen5Value =
 				noiseGen5Value > 0.0F
-					? (float)(this->noiseGen3.get()->GenerateOctaves((double)((float)blockX * 0.25714284F * 2.0F),
-																	 (double)((float)blockZ * 0.25714284F * 2.0F)) *
+					? float(this->noiseGen3.get()->GenerateOctaves(double((float)blockX * 0.25714284F * 2.0F),
+																	 double((float)blockZ * 0.25714284F * 2.0F)) *
 							  (double)noiseGen6Value / 4.0)
-					: (float)(this->noiseGen4.get()->GenerateOctaves((double)((float)blockX * 0.25714284F),
-																	 (double)((float)blockZ * 0.25714284F)) *
+					: float(this->noiseGen4.get()->GenerateOctaves(double((float)blockX * 0.25714284F),
+																	 double((float)blockZ * 0.25714284F)) *
 							  (double)noiseGen6Value);
 			int32_t terrainHeight = int32_t(noiseGen1Value + 64.0F + noiseGen5Value);
 			if ((float)this->noiseGen5.get()->GenerateOctaves((double)blockX, (double)blockZ) < 0.0F) {
 				terrainHeight = terrainHeight / 2 << 1;
-				if ((float)this->noiseGen5.get()->GenerateOctaves((double)(blockX / 5), (double)(blockZ / 5)) < 0.0F) {
+				if ((float)this->noiseGen5.get()->GenerateOctaves(double(blockX / 5), double(blockZ / 5)) < 0.0F) {
 					++terrainHeight;
 				}
 			}
